backend/requirements.txt
# =========================
# Framework principal
# =========================
fastapi==0.111.0
uvicorn[standard]==0.30.1

# =========================
# Base de datos (PostgreSQL)
# =========================
sqlalchemy==2.0.30
# Driver asíncrono (Recomendado para FastAPI)
asyncpg==0.29.0
# Driver síncrono (Opcional, útil si Alembic te da problemas con asyncpg)
psycopg2-binary==2.9.9 

# =========================
# Migraciones
# =========================
alembic==1.13.1

# =========================
# Autenticación y seguridad
# =========================
# MEJORA: PyJWT es el estándar moderno y mantenido
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# =========================
# Validación y configuración
# =========================
pydantic==2.7.1
# pydantic-settings ya maneja el archivo .env por ti
pydantic-settings==2.2.1
email-validator==2.1.1

# =========================
# Manejo de archivos / forms
# =========================
python-multipart==0.0.9

# =========================
# Generación de PDFs
# =========================
reportlab==4.2.0

# =========================
# Testing y Desarrollo (NUEVO)
# =========================
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.7

# =========================
# Utilidades
# =========================
python-slugify==8.0.4